requirements.txt
streamlit
fpdf2
python-dotenv
google-genai
pandas
psycopg2-binary